Full Update for HybridMcCormick Object
- Updates LPs, RR, probing for HybridMC object
- Separates load/optimize to allow for better inclusion of custom routines

+++ src/src/NLPSolver/src/Utilities/General.jl
@@ -3,7 +3,7 @@
 
 Checks that values `X1` and `X2` are equal to within tolerance `atol`.
 """
-function tol_eq(X1,X2,atol)
+function tol_eq(X1::Float64,X2::Float64,atol::Float64)
   out::Bool = true
   if (abs(X1-X2)>atol)
     out = false
